[WIP] Move profile trees into instrument
To add per-callstack function counts, we essentially require a tree structure.
That implies moving profile trees into the instrument. The new
implementation is relatively fast, but there is a bottleneck in the search
algorithm that determines how to go deeper into the tree.

In addition, this commit has no serialization of the resulting tree structure
that could be understood by the wrapper. It is a critical issue how to encode a
tree efficiently, or even what the wrapper would be doing with the trees after
it decodes them.