Count function calls inside the instrument
Since the current architecture follows every motion of the instrumentee's stack,
it allows us to count function calls in the instrument despite that the
interface between the instrument and wrapper is a sampler.

As of this commit, the instrument stores all function call counts in a tree.
When the stack is sampled, the sample includes the number of calls at
each callsite. If a counter gets sampled, it is cleared.  Counters which are not
sampled will continue to increment until they end up in a stack trace of some
function that is running while a sample is being taken.

This approach minimally changes the interface between the instrument and
wrapper. However, the semantics of the function call counts do not entirely
square with the supposed semantics of the profiles that the wrapper sends to the
backend; the call counts are not necessarily specific to a given profile period
(currently 1 minute). Functions that are rarely sampled will most likely have
counters that are have been accruing over more than one minute. The signficance
of this discrepancy is unclear.

It is known `hasCallee()` is a performance bottleneck.