Maintain strict profile semantics
To ensure that the numbers contained in a profile pertain to a particular time
window and not to incidental factors, profile construction and serialization
were moved completely into the instrument.

The instrument efficiently constructs profiles by tracking the program state
with a pointer to the current tree node (analogous to a stack pointer, but
saving previously-visited states).  It marshals trees to JSON, eliminating the
need to convert between different serializations in the wrapper (for instance,
s-expressions and JSON).  A significant amount of code has been added to
facilitate doing this robustly, avoiding the possibilities of buffer overruns
and system call interruption due to signal handlers.

The wrapper verifies the JSON output of the instrument and passes it to Kafka,
if connected.

It was found that the hasCallee() bottleneck previously identified was an
illusion, an artifact of compiling with gmon/gprof support. The benchmark `time
tree ~/source` takes 2s with gprof, 0.6-0.8 without. We conclude that the
performance of the profiler is acceptable.