Various profiler and test refinements
Use sarama.SyncProducer instead of AsyncProducer, which is simpler. See [1].

If the wrapper receives ^C, remove the socket before exiting. Ideally no matter
how the wrapper exits, the socket should be removed; unclear how to do this.  A
future commit should name the socket according to wrapper PID, so that
collisions are far less likely.

If the test instrumentee receives a negative command-line argument, run forever.
Useful for testing the connection to Kafka.

Add a network test case for the instrument.

[1]: https://talks.golang.org/2013/bestpractices.slide#25