Name sockets according to wrapper PID
This allows multiple wrapper instances to run in the same working directory
without their socket addresses colliding.  Plus, if a wrapper exits without
removing its socket, the next instance is unlikely to have the same pid,
reducing the probability of an address collision.

The instrument knows the PID of the process that forked it, so it uses the
correspondingly-named socket.