Profile program exit
The instrument emits a profile tree and frees memory upon the receipt of signals
for which the default behavior is termination. It also emits a profile and frees
memory at program shutdown via a destructor-attributed function. Technically the
instrumentee is not necessarily exited after our destructor runs, but it's
apparently none of the profiler's business how exactly the program shuts down.
Nevertheless, our destructor has to disable instrumentation to prevent it
running in user-defined destructors which have yet to be executed.