Implement error handling
According to its goals, the profiler should not interfere with user experience.
This commit pursues this goal by shutting down the instrument if anything goes
wrong. Error conditions are returned to callers, finally causing the instrument
to fallback to an offline state. Since failure can happen during initialization
or main execution, an array of pointers to configuration functions is used to
keep track of what changes in global state need to be reverted.