Add commit hash and top-level directory to release
A release is now associated with a Git commit, which is necessary for Git
integration. The system can use this information to generate source code
snippets of function definitions or callsites for the user interface.

DWARF line entries contain absolute paths of source files. It is possible that
one of these paths is not part of a Git repository: it could be an untracked
file or be located outside the repository. To detect this, the absolute path of
the top-level directory of the repository is now associated with a release.