Fix several instrument bugs
The fault injector had a bug in which a call to realloc() would succeed and move
a callee list to a new address, but the fault injector would present it as a
failure, so that the callee list pointer would not be updated. The program
segfaulted when the old address was passed to free(). This is fixed by ensuring
that realloc() is called only if a fault is not going to be injected.

The default sigaction for SIGPROF is termination. By reverting to this sigaction
upon error-triggered shutdown, the instrumentee can be terminated if SIGPROF is
raised, which occurs when ITIMER_PROF expires. I tried to prevent expiration by
stopping the timer, but this seemed not to work. Instead, the sigaction for
SIGPROF is now set to SIG_IGN; SIGPROF is always raised upon timer expiration,
but is ignored after error-triggered shutdown. The same goes for SIGVTALRM and
ITIMER_VIRTUAL.

The JSON buffer has been moved to global scope and its memory is now managed by
an enabler-disabler pair.

The error status of newNode() within appendCallee() is now handled.