[WIP] Use Makefiles for building and testing
Testing and building were previously afterthoughts in the organization of the
repo. They have become a main concern, and now it makes sense to take a more
systematic approach. Makefiles are used to perform most functions previously
done by scripts, and holistic tests have been consolidated and given a dedicted
location. This is a WIP because there are changes in the Makefiles that
anticipate future changes to the wrapper.

A top-level Makefile has been added to allow building, installation, and testing
of each component in the repository, replacing the utility scripts. Accordingly,
test scripts in component directories have been removed, and circle.yml now
invokes make.  A holistic test script exists in test/target/run.sh, and is
invoked by running `make test` at the top-level.

There used to be separate C programs for testing the releaser and the
instrument. They have been replaced by a single C program in test/src. The
Makefile for this program has new targets for installation and testing.

README.md's for each component have been updated to reflect the new building and
testing scheme.

All binaries are installed to appear in $PATH. The instrument is also installed
within the linker's default search path.

.gitignore has been updated to reflect the new directory structure.