[WIP] Fix CircleCI errors
It is too complicated to run the holistic test (`make test`) in CircleCI, since
it requires certs, which cannot be stored as files. A better place to run it is
in a client-side pre-commit hook. The use of a recursive Makefile in a hook
caused a strange bug in which the submake did not know it was running in a Git
repo (cause unknown). We decided to avoid recursion and use a single Makefile
for the whole repo.

POSIX and GNU source flags have been added when building the test program.