Change sample count behavior
The flamegraph used by the front-end assumes that the number of samples of a
node is greater than or equal to the sum of the number of samples of each child
node. To create this behavior, a sample now counts how many times a function was
on the stack _at all,_ not just at the _top_ of the stack. For each call to
`sig_sample()`, the nsamples counter of every functions currently on the stack
is incremented.