Rewrite wrapper
Some goroutines, that work on resources that need to be freed, have been wrapped
in a "launcher" that creates the resource, launches the goroutine, and returns a
cleanup function for the resource, which main can defer, ensuring that the
resource is freed no matter how the program terminates. This is a workaround to
the fact that a function deferred in a goroutine is not guaranteed to run when
main exits; there may be a better approach.

Exit events now include, timestamp, exit status, and signal (if the process was
signaled).

All Kafka messages contain a UUID and checksum.

SIGINT is relayed by the wrapper to the child, providing a way to stop execution
during manual testing. This also tests the exit event logic.

Log messages have been made more readable.