Use environment variables for configuration
All configurable parameters are now determined by environment variables. This
has yet to be tested against the backend.

If a variable has an empty value, wrap or release will report which variables
are missing and declare the configuration incomplete.

PEM certs are now expected to be base64-encoded.

The broker list must be a comma-delimited list of URLs.