Emit stacktrace on error signal
Certain signals should be regarded as indicating an error condition and are
therefore something the profiler should report. The developer, by linking to the
profiler library, decides to handle these signals by means of the profiler, not
by defining and installing their own signal handlers.

A separate socket is used exclusively for stacktrace JSON. This JSON includes a
signal number and an array of frames, including the root frame. The frames are
listed with the top of the stack first.