Use both realtime and cputime to time emissions
A profile tree is emitted if 1 second of CPU time or 10 seconds of realtime
have elapsed since the last emission. The test program demonstrates two
functions, one that uses 3 seconds of CPU time, and another 30 seconds of
realtime. We expect to see 3 profiles emitted due to SIGRT(REAL) and 2 due to
SIGRT(VIRT).