Refactor event handling
Since there are two goroutines that can produce Objects, both have to be wrapped
in another that is responsible for closing the channel when they exit.

Closures are used to wrap goroutines in a "synchronized box".

A lot of other changes have been made to the ends of organization, separation of
concerns, and readability, and more work remains.