Use type tags in instrument-wrapper interface
The schema is now the following:

	type string
	data json.RawMessage

This makes it easier to infer the type of objects being sent by the
instrument. By using json.RawMessage, we also avoid unmarshalling an
entire object before we know what to do with it. After the type is
extracted, we can convert unmarshal the data field into to an object
of the correct type.