wrap.go: Emit only one event
When the instrument handles an error signal and then calls _exit, the
OS does not consider the process to be "signaled" upon exit. That
status is reserved only for processes that do not handle the
terminating signal. This means that the wrapper has no way to tell
just by the exit status whether the process got an error signal.

A temporary fix is to not emit events for successful exit statuses.
The "correct" solution is to allow the cmd.Wait() part of the wrapper
to communicate with the Objectify() part so that successful exit
events can be generated.