Fix that bug
lib.c:

marshals is no longer responsible for conveying the exit status (since
it was faking it before anyway). The wrapper can do this itself.

wrap.go:

run has been moved into relay and Objectify. This allows Objectify to
wait for the child to exit, so that it can create the correct event.
If Objectify never receives an event, then relay's main loop will
eventually hit the socket EOF and a non-stacktrace event will be
generated.