wrap.go: Send wrapper logs to Kafka
By implementing a type, LogWriter, that satisfies io.Writer, we can
use the same log.Logger interface---previously used to write to
os.Stdout---to produce messages to Kafka.

A significant change is that because logs now send to the Kafka
producer, the producer itself cannot generate any logs while it is
consuming Objects, which would result in an infinite feedback loop.

Three log levels have been created.

The standard logger that writes to os.Stdout is still used for UI
when the configuration is incomplete or wrap is not called with any
arguments.